+++ android/app/src/debug/java/com/labfytools/trainlog/ui/DraftUiTestActivity.kt
@@ -98,7 +98,6 @@ class DraftUiTestActivity : ComponentActivity() {
                         onBody = {},
                         onOpenLatestSession = {},
                         onSync = {},
-                        onGenerateSession = {},
                     )
                 }
             }